requirements.txt
python-dotenv
discord.py
tortoise-orm
aiosqlite
python-dateutil
Pillow
python-slugify
